+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 [Unreleased]
 ### New Features
+
+### Fixes
+
+### Other
+
+## [0.0.9] - 2019-03-23
+### New Features
 - Added autocomplete support for LaTeX, HTML and markdown with preferences to disable each
 - Added support for WikiLink (`[[Description|./path/to/note.md]]`) markdown extensions
 - Added RTL writing direction support in editor and preview